Use security-key hint when authenticators are limited to roaming
There are broadly two types of roaming authenticators, security
keys and mobile phones. Hint to the browser to default to the
security key one when $wgWebAuthnLimitPasskeysToRoaming is enabled
as that was the purpose behind that switch. This is especially
useful since mobile phones don't actually work (T358771).

webauthn-lib doesn't support the 'hints' property so hack it in
on the JS side.

Bug: T399665
Change-Id: Iff0c07fd5bf854a0b8fff3b2458dabd0c22e6381
(cherry picked from commit 32b1c882de3803207187d8da65cd0825ad8908b1)